backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.2.34
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.15
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
